chore: moved menu buttons
